van tinh updated code for exercise 2
