--- README.md
@@ -1,7 +1,0 @@
-# Systems-Development-and-Frameworks
-
-| Github Alias    | Name         |
-| --------------- | ------------ |
-| @michaelef123           | Michael Franz     |
-| @Chuvantinh | Chu Văn Tình      |
-| @itfak     | Arnauld Feussi   |
