+++ VueJS/1/index.html
@@ -1,17 +1,17 @@
 <!-- development version, includes helpful console warnings -->
 <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
-
+<meta charset="UTF-8">
 <div id="hello">
 	<h2>{{ message }}</h2>
 </div>
 
 <div id="app">
-<todo-list
+<todos
 	v-bind:todos="todos"
 	v-on:save="save"
 	v-on:remove="remove"
 	v-on:add="add"
-></todo-list>
+></todos>
 </div>
 
-<script type="module" src="/index.js"></script>
+<script type="module" src="./index.js"></script>
