--- VueJS/README.md
@@ -1,17 +1,0 @@
-# Exercise \#1
-
-Implement a simple todo app with VueJS.
-
-Each item in the todo app should:
-
-* Edit: display a form to update the todo
-* Save: update the todo and display it
-* Cancel: cancel the form submission
-* Delete: delete the todo
-
-Here is how the result should look like:
-
-![screencast](./exercise-vuejs.gif)
-
-Use computed properties and events where possible and make use of the component
-composition pattern.
