--- VueJS/components/List.js
@@ -1,1 +1,0 @@
-export default Vue.component('list', {})
